attribute numByes in Downloader.kt and totalBytesSent in Uploader.kt changed to Long. numBytes in ClientResponse.kt changed to Long.

+++ libndt7/src/main/java/net/measurementlab/ndt7/android/models/ClientResponse.kt
@@ -10,5 +10,5 @@ data class ClientResponse(
 
 data class AppInfo(
     @SerializedName("ElapsedTime") val elapsedTime: Long,
-    @SerializedName("NumBytes") val numBytes: Double
+    @SerializedName("NumBytes") val numBytes: Long
 )
